+++ src/routes/PatronPreRegistrationRecordsContainer.test.js
@@ -34,6 +34,7 @@ jest.mock('react-router-dom', () => ({
   useHistory: jest.fn(),
 }));
 
+const STRIPES = buildStripes();
 const history = createMemoryHistory();
 history.push = jest.fn();
 const props = {
@@ -50,7 +51,10 @@ const props = {
     },
     resultOffset: 100
   },
-  stripes: buildStripes(),
+  stripes: {
+    ...STRIPES,
+    hasPerm: jest.fn().mockReturnValue(true),
+  },
   history,
 };
 
@@ -114,5 +118,21 @@ describe('PatronPreRegistrationRecordsContainer', () => {
     fireEvent.click(screen.getByTestId('need-more-button'));
 
     expect(mockSource.fetchMore).toHaveBeenCalledWith(100);
+  });
+
+  describe('when logged in user user do not have permission to view staging records', () => {
+    it('should render NoPermissionMessage component', () => {
+      const alteredProps = {
+        ...props,
+        stripes: {
+          ...props.stripes,
+          hasPerm: jest.fn().mockReturnValue(false),
+        },
+      };
+
+      renderPatronPreRegistrationRecordsContainer(alteredProps);
+
+      expect(screen.getByText('ui-users.stagingRecords.message.noAccessToStagingRecordsPage')).toBeDefined();
+    });
   });
 });