+++ CHANGELOG.md
@@ -73,6 +73,8 @@
 * Create new user record from preregistration data when no matching folio user is found. Refs UIU-3223.
 * Update permission name after Review and cleanup Module Descriptor for ui-checkout. Refs UIU-3216.
 * Integrate action to merge preregistration patron record with existing user record. Refs UIU-3225.
+* Update translations for pre registrations value. Refs UIU-3241.
+* Display a message to a user without permission on the “Patron preregistration record results” page. Refs UIU-3242.
 * Fix user edit without "Auth-Users" capability sets. Refs UIU-3243.
 
 ## [10.1.2](https://github.com/folio-org/ui-users/tree/v10.1.2) (2024-09-05)